Add logic for degeneracy prevention

+++ build_task.go
@@ -8,15 +8,15 @@ type Task struct {
 }
 
 type tableOuterCell struct {
-	amount    int
-	potential int
+	amount    float64
+	potential float64
 	isFake    bool
 }
 
 type tableCell struct {
-	cost           int
-	deliveryAmount int
-	delta          int
+	cost           float64
+	deliveryAmount float64
+	delta          float64
 	sign           rune
 }
 
@@ -25,12 +25,12 @@ func buildTaskFromParams(params Params) Task {
 
 	task.supplyList = make([]tableOuterCell, len(params.SupplyList))
 	for i, val := range params.SupplyList {
-		task.supplyList[i] = tableOuterCell{amount: val}
+		task.supplyList[i] = tableOuterCell{amount: float64(val)}
 	}
 
 	task.demandList = make([]tableOuterCell, len(params.DemandList))
 	for i, val := range params.DemandList {
-		task.demandList[i] = tableOuterCell{amount: val}
+		task.demandList[i] = tableOuterCell{amount: float64(val)}
 	}
 
 	task.tableCells = make([][]tableCell, len(params.SupplyList))
@@ -38,7 +38,7 @@ func buildTaskFromParams(params Params) Task {
 		// assign table row
 		task.tableCells[i] = make([]tableCell, len(row))
 		for j, cost := range row {
-			task.tableCells[i][j] = tableCell{cost: cost}
+			task.tableCells[i][j] = tableCell{cost: float64(cost)}
 		}
 	}
