+++ north_west_corner.go
@@ -4,8 +4,8 @@ func (t *Task) northWestCorner() {
 	u := 0 // supplier index
 	v := 0 // demand index
 	// already supllied sums
-	aS := make([]int, len(t.demandList))
-	aD := make([]int, len(t.supplyList))
+	aS := make([]float64, len(t.demandList))
+	aD := make([]float64, len(t.supplyList))
 
 	for u < len(t.supplyList) && v < len(t.demandList) {
 		if t.demandList[v].amount-aS[v] < t.supplyList[u].amount-aD[u] {
